:wrench: Move partners page  within new pages
