+++ src/controllers/admin/index.ts
@@ -8,8 +8,7 @@ export * from './getPeriodesCsv'
 export * from './getRegenerateCertificatesPage'
 export * from './getSignalerDemandeAbandonPage'
 export * from './getSignalerDemandeDelaiPage'
-export * from './getUsersPage'
-export * from './postAppelsOffres'
+export * from './getPartnersListPage'
 export * from './postInviteDrealUser'
 export * from './postInvitePartnerUsers'
 export * from './postPeriodes'
