+++ src/views/legacy-pages/index.ts
@@ -4,7 +4,6 @@ import { makeHtml } from '../index.html'
 import AdminAppelOffre from './adminAppelOffre'
 import AdminNotifyCandidates from './adminNotifyCandidates'
 import AdminRegenerateCertificates from './adminRegenerateCertificates'
-import AdminUsers from './adminUsers'
 import ImportCandidates from './importCandidates'
 import InvitationList from './invitationList'
 import NotificationList from './notificationList'
@@ -22,7 +21,6 @@ const InvitationListPage = makePresenterPage(InvitationList)
 const NotificationListPage = makePresenterPage(NotificationList)
 const SuccessOrErrorPage = makePresenterPage(SuccessOrError)
 const AdminAppelOffrePage = makePresenterPage(AdminAppelOffre)
-const AdminUsersPage = makePresenterPage(AdminUsers)
 const ListMissingOwnerProjectsPage = makePresenterPage(ListMissingOwnerProjects)
 const InvitationsAreDeprecatedPage = makePresenterPage(InvitationsAreDeprecated)
 const FakeLoginPage = makePresenterPage(FakeLogin)
@@ -35,7 +33,6 @@ export {
   NotificationListPage,
   SuccessOrErrorPage as SuccessPage,
   AdminAppelOffrePage,
-  AdminUsersPage,
   AdminRegenerateCertificatesPage,
   ListMissingOwnerProjectsPage,
   InvitationsAreDeprecatedPage,
