+++ src/views/pages/index.ts
@@ -4,6 +4,7 @@ export * from './homePage'
 export * from './modificationRequestListPage'
 export * from './modificationRequestPage'
 export * from './newModificationRequestPage'
+export * from './partnersListPage'
 export * from './projectDetailsPage'
 export * from './signalerDemandeAbandonPage'
 export * from './signalerDemandeDelaiPage'
