+++ src/views/pages/partnersListPage/index.ts
@@ -1,0 +1,1 @@
+export * from './PartnersList';
